refactor(process): split python.rs into per-class PyO3 binding siblings
python.rs (973 lines) becomes a 69-line thin facade that mounts seven focused siblings via #[path]: python_helpers (shared Tokio runtime + error/status adaptors), python_monitor, python_launcher, python_crash_policy (+ parse_status), python_watcher (+ internal PyWatcherEvent), python_standalone_dispatcher, and python_pumped_dispatcher (+ parse_affinity / outcome_to_dict). register_classes stays in the facade as the single PyModule registration point; every Py* class is re-exported so downstream call sites keep using crate::python::* unchanged.

Drive-by: convert the rst-style 'Example::' blocks in each Py* class docstring to fenced python code blocks so they no longer fail rustdoc parsing when --features python-bindings --doc is run (the 5 failures were already latent on HEAD and surface only when the feature flag is enabled).

84 lib tests + 1 doctest green with --features python-bindings; workspace cargo test --workspace stays green; docs/guide/architecture.md gains a Maintainer layout section under dcc-mcp-process.
